refactor(vm): split vm_misc_ops.rs into cohesive modules (§7-8)
Split the 3600-line vm_misc_ops.rs (over the 500-line limit) into ten
cohesive modules, all under 500 lines:

- vm_misc_ops.rs (377)            - ReductionAssoc, core-type helpers,
                                    reduction-callable setup
- vm_misc_typed_range.rs (346)    - typed-array constraints + Range ops
- vm_misc_coerce.rs (404)         - num/str coercion + pre-incr/decr
- vm_misc_codevar.rs (251)        - code/symbolic-var + env-index ops
- vm_misc_assign.rs (484)         - assign-expr inner + native-int validation
- vm_misc_reduction_exec.rs (479) - exec_reduction_op
- vm_misc_reduction_scan.rs (313) - scan reductions + magic/scope ops
- vm_misc_typecheck.rs (288)      - type-check op inner
- vm_misc_scope.rs (375)          - state-var init + block-scope op
- vm_misc_block.rs (313)          - do/once/let block exprs + check-phaser

All methods are impl Interpreter, so the split is visibility-only. The
ReductionAssoc enum + core-type free fns stay in the main module (bumped
to pub(super), glob-imported where used); cross-module private helpers
are bumped to pub(super). The fn multiset is byte-identical. No behavior
change.
