+++ src/vm.rs
@@ -151,7 +151,16 @@ pub(crate) mod vm_hyper_ops;
 mod vm_hyper_race_parallel;
 pub(crate) mod vm_meta_ops;
 pub(crate) mod vm_method_dispatch;
+pub(crate) mod vm_misc_assign;
+pub(crate) mod vm_misc_block;
+pub(crate) mod vm_misc_codevar;
+pub(crate) mod vm_misc_coerce;
 pub(crate) mod vm_misc_ops;
+pub(crate) mod vm_misc_reduction_exec;
+pub(crate) mod vm_misc_reduction_scan;
+pub(crate) mod vm_misc_scope;
+pub(crate) mod vm_misc_typecheck;
+pub(crate) mod vm_misc_typed_range;
 mod vm_mixin_does_ops;
 mod vm_module_ops;
 mod vm_native_dispatch;
